// lib/src/types.ts
export interface BaseData {
  id?: string | null;
  _id?: string | null;
  [key: string]: any;
}

export type PathType<RowType extends BaseData> = Exclude<LiteralUnion<keyof RowType, string>, symbol | number>;

export type PathValueType<T extends BaseData> = PathType<T> | true | undefined;

export type Sorter<T> = (ab: T, ba: T) => number;

export type TableCellAlign = "left" | "center" | "right";
// ...
type WithDataIndex<RowType extends BaseData, AllDataType extends RowType[]> = BaseColumnDefinition<
  RowType,
  AllDataType
> & {
  dataIndex: PathType<RowType>;
};

type WithExactlyOne<
  RowType extends BaseData,
  AllDataType extends RowType[],
  Requires extends keyof BaseColumnDefinition<RowType, AllDataType> = "numerical" | "render",
> = RequireExactlyOne<BaseColumnDefinition<RowType, AllDataType>, Requires>;

type GetColumnDefinition<RowType extends BaseData, AllDataType extends RowType[] = RowType[]> =
  | WithDataIndex<RowType, AllDataType>
  | WithExactlyOne<RowType, AllDataType, "numerical" | "render" | "colGroup" | "hidden">;

export type ColumnDefinition<RowType extends BaseData, AllDataType extends RowType[] = RowType[]> = GetColumnDefinition<
  RowType,
  AllDataType
>;

export type FullColDef<RowType extends BaseData, AllDataType extends RowType[]> = ColumnDefinition<
  RowType,
  AllDataType
> & {
  /**
   * @private internal use only
   */
  parentKey?: never;
  /**
   * @private internal use only
   */
  isColGroup?: false;
  /**
   * @private internal use only
   */
  hasColGroupFooter?: never;
};

export type ColGroupDefinition<RowType extends BaseData, AllDataType extends RowType[] = RowType[]> = Omit<
  WithDataIndex<RowType, AllDataType> | WithExactlyOne<RowType, AllDataType>,
  "colGroup"
> & { colGroup?: never };

export type FullColGroupDef<RowType extends BaseData, AllDataType extends RowType[]> = ColGroupDefinition<
  RowType,
  AllDataType
> & {
  /**
   * @private internal use only
   */
  parentKey?: string;
  /**
   * @private internal use only
   */
  isColGroup?: true;
  /**
   * @private internal use only
   */
  hasColGroupFooter?: boolean;
};
// ...
export interface ActionButton extends Omit<IconButtonProps, "size"> {
  key: string;
  icon: ReactNode;
  onClick(): void;
}
// ...
export interface Sort {
  key: null | string;
  direction?: "asc" | "desc";
}